+++ TelnetNegotiationCore.UnitTests/GMCPTests.cs
@@ -55,7 +55,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Server",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }
@@ -128,7 +128,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Client",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }
@@ -251,7 +251,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Server",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }
@@ -336,7 +336,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Client",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }
@@ -421,7 +421,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Server",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }
@@ -506,7 +506,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Client",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }
@@ -562,7 +562,7 @@ public class GMCPTests : BaseTest
 			Name = "My Telnet Negotiated Server",
 			UTF_8 = true,
 			Gameplay = ["ABC", "DEF"],
-			Extended = new Dictionary<string, dynamic>
+			Extended = new Dictionary<string, object>
 			{
 				{ "Foo", "Bar"},
 				{ "Baz", (string[]) ["Moo", "Meow"] }